requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
evdev==1.9.2
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.3.1
pillow==11.3.0
Pygments==2.19.2
pynput==1.8.1
python-xlib==0.33
requests==2.32.4
rich==14.0.0
six==1.17.0
urllib3==2.5.0
